requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
django-storages[google]==1.12.3
django-environ==0.8.1
